Got project working by changing properties, and moved image to drawable
