+++ AudiOcean/AudiOcean/AudiOcean/HomePage.xaml.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
-namespace GUILibrary
+namespace AudiOcean
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class HomePage : ContentPage
