+++ AudiOcean/AudiOceanServer/packages.config
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Carbon.Forms" version="1.5.0" targetFramework="net47" />
-  <package id="Carbon.Json" version="1.5.0" targetFramework="net47" />
-  <package id="Carbon.OAuth2" version="0.11.0" targetFramework="net47" />
+  <package id="Carbon.Forms" version="1.6.0" targetFramework="net47" />
+  <package id="Carbon.Json" version="1.6.1" targetFramework="net47" />
+  <package id="Carbon.OAuth2" version="0.14.0" targetFramework="net47" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net47" />
   <package id="Google.Authentication" version="0.17.0" targetFramework="net47" />
   <package id="System.Text.Encodings.Web" version="4.4.0" targetFramework="net47" />
